+++ src/multipac_testbench/multipactor_test.py
@@ -18,7 +18,7 @@
 import itertools
 import logging
 from abc import ABCMeta
-from collections.abc import Iterable, Sequence
+from collections.abc import Collection, Iterable, Sequence
 from pathlib import Path
 from typing import Any
 
@@ -65,7 +65,6 @@ class MultipactorTest:
         swr: float,
         info: str = "",
         sep: str = "\t",
-        verbose: bool = False,
         **kwargs,
     ) -> None:
         r"""Create all the pick-ups.
@@ -84,9 +83,6 @@ class MultipactorTest:
             An additional string to identify this test in plots.
         sep :
             Delimiter between two columns in ``filepath``.
-        verbose :
-            To print information on the structure of the test bench, as it was
-            understood.
 
         """
         self.filepath = filepath
@@ -118,7 +114,6 @@ class MultipactorTest:
         imeasurement_points = imeasurement_point_factory.run(
             config,
             df_data,
-            verbose,
         )
         #: Where all diagnostics at a specific pick-up are defined (e.g.
         #: current probe)
@@ -139,19 +134,34 @@ class MultipactorTest:
         return ", ".join(out)
 
     def add_post_treater(
-        self, *args, only_pick_up_which_name_is: tuple[str, ...] = (), **kwargs
+        self, *args, only_pick_up_which_name_is: Collection[str] = (), **kwargs
     ) -> None:
-        """Add post-treatment functions to instruments."""
-        pick_ups = self.pick_ups
+        """Add post-treatment functions to instruments.
+
+        .. todo::
+            Find out why following lines result in strange plot linestyles.
+
+            .. code-block:: py
+
+                measurement_points: list[IMeasurementPoint] = self.pick_ups
+                if self.global_diagnostics is not None:
+                    measurement_points.append(self.global_diagnostics)
+
+
+        """
+        measurement_points: list[IMeasurementPoint] = self.pick_ups
+        if self.global_diagnostics is not None:
+            measurement_points = self.pick_ups + [self.global_diagnostics]
+
         if len(only_pick_up_which_name_is) > 0:
-            pick_ups = [
-                pick_up
-                for pick_up in self.pick_ups
-                if pick_up.name in only_pick_up_which_name_is
+            measurement_points = [
+                point
+                for point in measurement_points
+                if point.name in only_pick_up_which_name_is
             ]
 
-        for pick_up in pick_ups:
-            pick_up.add_post_treater(*args, **kwargs)
+        for point in measurement_points:
+            point.add_post_treater(*args, **kwargs)
 
     def sweet_plot(
         self,